+++ .eslintignore
@@ -1,4 +1,4 @@
-src/invoke-lambda/runtimes/deno.js
-src/invoke-lambda/runtimes/node-esm.js
+src/invoke-lambda/exec/runtimes/deno.js
+src/invoke-lambda/exec/runtimes/node-esm.js
 test/mock
 scratch/
